+++ src/client/tests/ConnectionTest.cpp
@@ -3,12 +3,12 @@
  * This source code is licensed under Apache 2.0 License.
  */
 
+#include <common/Init.h>
 #include <folly/json.h>
 #include <folly/synchronization/Baton.h>
 #include <glog/logging.h>
 #include <gtest/gtest.h>
 #include <nebula/client/Connection.h>
-#include <nebula/client/Init.h>
 
 #include "../../ClientTest.h"
