+++ src/client/tests/SessionTest.cpp
@@ -3,6 +3,7 @@
  * This source code is licensed under Apache 2.0 License.
  */
 
+#include <common/Init.h>
 #include <common/datatypes/Geography.h>
 #include <folly/json.h>
 #include <folly/synchronization/Baton.h>
@@ -11,7 +12,6 @@
 #include <gtest/gtest_prod.h>
 #include <nebula/client/Config.h>
 #include <nebula/client/ConnectionPool.h>
-#include <nebula/client/Init.h>
 #include <nebula/client/Session.h>
 
 #include <atomic>
